requirements.txt
python-chess
requests
stockfish
beautifulsoup4